move well known cloud provider taints to k8s.io/cloud-provider/api
