+++ pkg/scheduler/BUILD
@@ -117,7 +117,6 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//pkg/scheduler/algorithmprovider:all-srcs",
-        "//pkg/scheduler/api:all-srcs",
         "//pkg/scheduler/apis/config:all-srcs",
         "//pkg/scheduler/apis/extender/v1:all-srcs",
         "//pkg/scheduler/core:all-srcs",
